requirements.txt
pip
python-binance